backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Abhinav Gupta"]
license = "MIT"
packages = [{ include = "apis" }]

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.45"
copilotkit = "^0.1.27"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
xmltodict = "^0.14.2"
arxiv = "^2.1.3"
pinecone = "^5.3.1"
langchain-community = "^0.3.5"
langchain = "^0.3.7"
torch = "^2.5.1"
sentence-transformers = "^3.3.0"
langchain-pinecone = {version = "^0.2.0", python = "<3.13"}
pinecone-client = "^5.0.1"
langchain-huggingface = "^0.1.2"
tavily-python = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
